// src/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
// ...
  public: {
// ...
    Views: {
      cities_with_categories: {
        Row: {
          category_code: string | null;
          category_description: string | null;
          category_id: string | null;
          category_name: string | null;
          country: string | null;
          cover_image: string | null;
          description: string | null;
          id: string | null;
          location: unknown;
          name: string | null;
        };
        Relationships: [];
      };
      cities_with_full_info: {
        Row: {
          categories: Json | null;
          country: string | null;
          cover_image: string | null;
          description: string | null;
          id: string | null;
          latitude: number | null;
          longitude: number | null;
          name: string | null;
          tourist_attractions: Json | null;
        };
        Insert: {
          categories?: never;
          country?: string | null;
          cover_image?: string | null;
          description?: string | null;
          id?: string | null;
          latitude?: never;
          longitude?: never;
          name?: string | null;
          tourist_attractions?: never;
        };
        Update: {
          categories?: never;
          country?: string | null;
          cover_image?: string | null;
          description?: string | null;
          id?: string | null;
          latitude?: never;
          longitude?: never;
          name?: string | null;
          tourist_attractions?: never;
        };
        Relationships: [];
      };
      geography_columns: {
        Row: {
          coord_dimension: number | null;
          f_geography_column: unknown;
          f_table_catalog: unknown;
          f_table_name: unknown;
          f_table_schema: unknown;
          srid: number | null;
          type: string | null;
        };
        Relationships: [];
      };
      geometry_columns: {
        Row: {
          coord_dimension: number | null;
          f_geometry_column: unknown;
          f_table_catalog: string | null;
          f_table_name: unknown;
          f_table_schema: unknown;
          srid: number | null;
          type: string | null;
        };
        Insert: {
          coord_dimension?: number | null;
          f_geometry_column?: unknown;
          f_table_catalog?: string | null;
          f_table_name?: unknown;
          f_table_schema?: unknown;
          srid?: number | null;
          type?: string | null;
        };
        Update: {
          coord_dimension?: number | null;
          f_geometry_column?: unknown;
          f_table_catalog?: string | null;
          f_table_name?: unknown;
          f_table_schema?: unknown;
          srid?: number | null;
          type?: string | null;
        };
        Relationships: [];
      };
      related_cities: {
        Row: {
          country: string | null;
          cover_image: string | null;
          id: string | null;
          name: string | null;
          source_city_id: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "city_cities_city_id_fkey";
            columns: ["source_city_id"];
            isOneToOne: false;
            referencedRelation: "cities";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "city_cities_city_id_fkey";
            columns: ["source_city_id"];
            isOneToOne: false;
            referencedRelation: "cities_with_categories";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "city_cities_city_id_fkey";
            columns: ["source_city_id"];
            isOneToOne: false;
            referencedRelation: "cities_with_full_info";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "city_cities_city_id_fkey";
            columns: ["source_city_id"];
            isOneToOne: false;
            referencedRelation: "related_cities";
            referencedColumns: ["id"];
          },
        ];
      };
    };
// ...
  };
};